+++ CHANGELOG.md
@@ -1,5 +1,8 @@
+# Changelog
+
 ### Unreleased (`canary` & `dev` dist-tags)
 
+ * Add fileWriteStreamHandler option
  * Add allowEmptyFiles and minFileSize options
  * Test only on Node.js >= v10. Support only Node LTS and latest ([#515](https://github.com/node-formidable/node-formidable/pull/515))
  * stop using deprecated features ([#516](https://github.com/node-formidable/node-formidable/pull/516), [#472](https://github.com/node-formidable/node-formidable/issues/472), [#406](https://github.com/node-formidable/node-formidable/issues/406))